chore(libs): remove circular dependencies from walletsync
